% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_island_colonist.R
\name{add_island_colonist}
\alias{add_island_colonist}
\title{Adds an island colonists (can be either a singleton lineage or an island
clade) to the island community (island_tbl).}
\usage{
add_island_colonist(
  island_tbl,
  clade_name,
  status,
  missing_species,
  branching_times,
  min_age
)
}
\arguments{
\item{island_tbl}{An instance of the `Island_tbl` class.}

\item{clade_name}{Character name of the colonising clade.}

\item{status}{Character endemicity status of the colonising clade.}

\item{missing_species}{Numeric number of missing species from the phylogeny
that belong to the colonising clade.}

\item{branching_times}{Numeric vector of one or more elements where the first
element is the colonisation time and subsequent elements are the branching
times on the island.}

\item{min_age}{Numeric minimum age (time before the present) that the species
must have colonised the island by. This is known when there is a branching
on the island, either in species or subspecies.}
}
\value{
An object of `Island_tbl` class
}
\description{
Adds an island colonists (can be either a singleton lineage or an island
clade) to the island community (island_tbl).
}
\examples{
island_colonist <- DAISIEprep::island_colonist(
  clade_name = "bird",
  status = "endemic",
  missing_species = 0,
  branching_times = c(1, 0.5)
)
island_tbl <- island_tbl()
bind_colonist_to_tbl(
  island_colonist = island_colonist,
  island_tbl = island_tbl
)
island_tbl <- add_island_colonist(
 island_tbl,
 clade_name = "new_clade",
 status = "endemic",
 missing_species = 0,
 branching_times = 1,
 min_age = NA
)
}
