% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_species.R
\name{add_missing_species}
\alias{add_missing_species}
\title{Adds the missing species to the clades that have missing species from the
tree in the Island_tbl object. This is to be used after
extract_island_species to input missing species.}
\usage{
add_missing_species(island_tbl, missing_species_df)
}
\arguments{
\item{island_tbl}{An instance of the `Island_tbl` class.}

\item{missing_species_df}{A data frame containing the names of the clades
extracted from the phylogeny and the number of missing species in each clade.
Not every clade extracted from the phylogeny needs to be included if the
number of missing species in them clades is zero.}
}
\value{
Data frame with single column of character strings and row names
}
\description{
Adds the missing species to the clades that have missing species from the
tree in the Island_tbl object. This is to be used after
extract_island_species to input missing species.
}
\examples{
missing_species <- data.frame(clade_name = "bird_c", missing_species = 1)
set.seed(
  1,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
phylo <- ape::rcoal(5)
phylo$tip.label <- c("bird_a", "bird_b", "bird_c", "bird_d", "bird_e")
phylo <- phylobase::phylo4(phylo)
endemicity_status <- c(
  "not_present", "not_present", "endemic", "not_present", "not_present"
)
phylod <- phylobase::phylo4d(phylo, as.data.frame(endemicity_status))
island_tbl <- extract_island_species(phylod, extraction_method = "min")
island_tbl <- add_missing_species(island_tbl, missing_species)
}
