% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outgroup.R
\name{add_outgroup}
\alias{add_outgroup}
\title{Add an outgroup species to a given phylogeny.}
\usage{
add_outgroup(phylo)
}
\arguments{
\item{phylo}{A phylogeny either as a \code{phylo} (from the \code{ape} package) or
\code{phylo4} (from the \code{phylobase} package) object.}
}
\value{
A \code{phylo} object
}
\description{
Add an outgroup species to a given phylogeny.
}
\examples{
phylo <- ape::rcoal(10)
phylo_with_outgroup <- add_outgroup(phylo)
}
