% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_back_colonisation.R
\name{any_back_colonisation}
\alias{any_back_colonisation}
\title{Detects any cases where a non-endemic species or species not present on the
island has likely been on the island given its ancestral state reconstruction
indicating ancestral presence on the island and so is likely a back
colonisation from the island to the mainland (or potentially different
island). This function is useful if using extraction_method = "min" in
DAISIEprep::extract_island_species() as it may brake up a single colonist
into multiple colonists because of back-colonisation.}
\usage{
any_back_colonisation(phylod)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}
}
\value{
A named vector
}
\description{
Detects any cases where a non-endemic species or species not present on the
island has likely been on the island given its ancestral state reconstruction
indicating ancestral presence on the island and so is likely a back
colonisation from the island to the mainland (or potentially different
island). This function is useful if using extraction_method = "min" in
DAISIEprep::extract_island_species() as it may brake up a single colonist
into multiple colonists because of back-colonisation.
}
