% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Performance analysis of the extract_island_species() function
Uses system.time() for timing for reasons explained here:
https://radfordneal.wordpress.com/2014/02/02/inaccurate-results-from-microbenchmark/}
\usage{
benchmark(
  tree_size_range,
  num_points,
  prob_on_island,
  prob_endemic,
  replicates,
  extraction_method,
  asr_method,
  tie_preference,
  log_scale = TRUE
)
}
\arguments{
\item{tree_size_range}{stub}

\item{num_points}{stub}

\item{prob_on_island}{stub}

\item{prob_endemic}{stub}

\item{replicates}{stub}

\item{extraction_method}{stub}

\item{asr_method}{stub}

\item{tie_preference}{stub}

\item{log_scale}{stub}
}
\value{
Data frame
}
\description{
Performance analysis of the extract_island_species() function
Uses system.time() for timing for reasons explained here:
https://radfordneal.wordpress.com/2014/02/02/inaccurate-results-from-microbenchmark/
}
\examples{
\dontrun{
#WIP
}
}
