% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_colonist.R
\name{check_duplicate_colonist}
\alias{check_duplicate_colonist}
\title{Check if colonist has already been stored in `Island_tbl` class. This is used
to stop endemic clades from being stored multiple times in the island table
by checking if the endemicity status and branching times are identical.}
\usage{
check_duplicate_colonist(island_colonist, island_tbl)
}
\arguments{
\item{island_colonist}{An instance of the `Island_colonist` class.}

\item{island_tbl}{An instance of the `Island_tbl` class.}
}
\value{
Boolean
}
\description{
Check if colonist has already been stored in `Island_tbl` class. This is used
to stop endemic clades from being stored multiple times in the island table
by checking if the endemicity status and branching times are identical.
}
\examples{
island_col <- island_colonist(
  clade_name = "bird",
  status = "endemic",
  missing_species = 0,
  branching_times = c(1.0, 0.5)
)
island_tbl <- methods::new("island_tbl")
check_duplicate_colonist(
  island_colonist = island_col,
  island_tbl = island_tbl
)
}
