% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_island_tbl-class.R
\name{check_multi_island_tbl}
\alias{check_multi_island_tbl}
\title{Checks the validity of the Multi_island_tbl class}
\usage{
check_multi_island_tbl(object)
}
\arguments{
\item{object}{Instance of the Multi_island_tbl class}
}
\value{
Boolean or errors
}
\description{
Checks the validity of the Multi_island_tbl class
}
\examples{
multi_island_tbl <- multi_island_tbl()
check_multi_island_tbl(multi_island_tbl)
}
