% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coccyzus_phylod}
\alias{coccyzus_phylod}
\title{A phylogenetic tree of coccyzus species with endemicity status as tip states.}
\format{
\subsection{\code{coccyzus_phylod}}{

A \code{phylo4d} object (from the \pkg{phylobase} package) with
20 tips and
19 internal
nodes.
}
}
\source{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015). Equilibrium
and non-equilibrium dynamics simultaneously operate in the Galapagos
islands. Ecology Letters 18: 844-852.
}
\usage{
coccyzus_phylod
}
\description{
A phylogenetic tree of coccyzus species with endemicity status as tip states.
}
\keyword{datasets}
