% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing_species.R
\name{count_missing_species}
\alias{count_missing_species}
\title{Reads in the checklist of all species on an island, including those that are
not in the phylogeny (represented by NA) and counts the number of species
missing from the phylogeny each genus}
\usage{
count_missing_species(
  checklist,
  in_phylo_col,
  endemicity_status_col,
  rm_species_col = NULL
)
}
\value{
Data frame
}
\description{
Reads in the checklist of all species on an island, including those that are
not in the phylogeny (represented by NA) and counts the number of species
missing from the phylogeny each genus
}
\examples{
mammal_checklist <- read_checklist(file_name = "mammal_checklist.csv")
count_missing_species(checklist = mammal_checklist)
}
