% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_endemicity_status.R
\name{create_endemicity_status}
\alias{create_endemicity_status}
\title{Creates a data frame with the endemicity status (either 'endemic',
'nonendemic', 'not_present') of every species in the phylogeny using a
phylogeny and a data frame of the island species and their endemicity (either
'endemic' or 'nonendemic') provided.}
\usage{
create_endemicity_status(phylo, island_species)
}
\value{
Data frame with single column of character strings and row names
}
\description{
Creates a data frame with the endemicity status (either 'endemic',
'nonendemic', 'not_present') of every species in the phylogeny using a
phylogeny and a data frame of the island species and their endemicity (either
'endemic' or 'nonendemic') provided.
}
\examples{
set.seed(1)
phylo <- ape::rcoal(4)
phylo$tip.label <- c("species_a", "species_b", "species_c", "species_d")
phylo <- as(phylo, "phylo4")
island_species <- data.frame(
  tip_labels = c("species_a", "species_b", "species_c", "species_d"),
  tip_endemicity_status = c("endemic", "endemic", "endemic", "nonendemic")
)
endemicity_status <- create_endemicity_status(
  phylo = phylo,
  island_species = island_species
)
}
