% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_phylod.R
\name{create_test_phylod}
\alias{create_test_phylod}
\title{Creates phylod objects.}
\usage{
create_test_phylod(test_scenario)
}
\arguments{
\item{test_scenario}{Integer specifying which test phylod object to create.}
}
\value{
A `phylo4d` object
}
\description{
A helper function that is useful in tests and examples to
easily create `phylod` objects (i.e. phylogenetic trees with data).
}
\examples{
create_test_phylod(test_scenario = 1)
}
