% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIEprep-package.R
\docType{package}
\name{DAISIEprep-package}
\alias{DAISIEprep}
\alias{DAISIEprep-package}
\title{DAISIEprep: Extracts Phylogenetic Island Community Data from Phylogenetic Trees}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extracts colonisation and branching times of island species to be used for analysis in the R package 'DAISIE'. It uses phylogenetic and endemicity data to extract the separate island colonists and store them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joshwlambert/DAISIEprep}
  \item \url{https://joshwlambert.github.io/DAISIEprep/}
  \item Report bugs at \url{https://github.com/joshwlambert/DAISIEprep/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{j.w.l.lambert@rug.nl} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID})

Authors:
\itemize{
  \item Luis Valente \email{luis.valente@naturalis.nl} (\href{https://orcid.org/0000-0003-4247-8785}{ORCID})
  \item Pedro Santos Neves \email{p.m.santos.neves@rug.nl} (\href{https://orcid.org/0000-0003-2561-4677}{ORCID})
  \item Lizzie Roeble \email{lizzie.roeble@naturalis.nl} (\href{https://orcid.org/0000-0003-3664-4222}{ORCID})
  \item Theo Pannetier \email{t.s.c.pannetier@rug.nl} (\href{https://orcid.org/0000-0002-8424-3573}{ORCID})
}

}
\keyword{internal}
