% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_param_doc}
\alias{default_param_doc}
\title{Documentation for DAISIEprep package}
\usage{
default_param_doc(island_colonist, island_tbl, phylod, species_label)
}
\arguments{
\item{island_colonist}{An instance of the `island_colonist` class}

\item{island_tbl}{An instance of the `island_tbl` class}

\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}
}
\value{
Nothing
}
\description{
Documentation for DAISIEprep package
}
