% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for function in the DAISIEprep package}
\usage{
default_params_doc(
  island_colonist,
  island_tbl,
  phylod,
  extraction_method,
  species_label,
  x,
  value,
  clade_name,
  status,
  missing_species,
  branching_times,
  min_age,
  endemic_clade,
  phylo,
  island_species
)
}
\arguments{
\item{island_colonist}{An instance of the `Island_colonist` class.}

\item{island_tbl}{An instance of the `Island_tbl` class.}

\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{extraction_method}{A character string specifying whether the
colonisation time extracted is the minimum time (`min`) (before the present),
or the most probable time under ancestral state reconstruction (`asr`).}

\item{species_label}{The tip label of the species of interest.}

\item{x}{An object whose class is determined by the signature.}

\item{value}{A value which can take several forms to be assigned to an object
of a class.}

\item{clade_name}{Character name of the colonising clade.}

\item{status}{Character endemicity status of the colonising clade.}

\item{missing_species}{Numeric number of missing species from the phylogeny
that belong to the colonising clade.}

\item{branching_times}{Numeric vector of one or more elements where the first
element is the colonisation time and subsequent elements are the branching
times on the island.}

\item{min_age}{Numeric minimum age (time before the present) that the species
must have colonised the island by. This is known when there is a branching
on the island, either in species or subspecies.}

\item{endemic_clade}{Named vector with all the species from a clade.}

\item{phylo}{A `phylo4` object from the package `phylobase`.}

\item{island_species}{Data frame with two columns. The first is a character
string of the tip_labels with the tip names of the species on the island.
The second column a character string of the endemicity status of the species,
either endemic or nonendemic.}
}
\value{
Nothing
}
\description{
Documentation for function in the DAISIEprep package
}
\author{
Joshua W. Lambert
}
