% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemicity_to_states.R
\name{endemicity_to_states}
\alias{endemicity_to_states}
\title{Convert endemicity to SSE states}
\usage{
endemicity_to_states(endemicity_status)
}
\arguments{
\item{endemicity_status}{character vector with values "endemic", "nonendemic"
and/or "not_present"}
}
\value{
an integer vector of tip states, as expected by SSE models
}
\description{
Convert endemicity to SSE states
}
