% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_biogeobears_input.R
\name{extract_biogeobears_ancestral_states_probs}
\alias{extract_biogeobears_ancestral_states_probs}
\title{Extract ancestral state probabilities from BioGeoBEARS output}
\usage{
extract_biogeobears_ancestral_states_probs(biogeobears_res)
}
\arguments{
\item{biogeobears_res}{a list, the output of \code{\link[BioGeoBEARS:bears_optim_run]{BioGeoBEARS::bears_optim_run()}}}
}
\value{
a data.frame with one row per node (tips and internals) and four
columns: label | not_present | endemic | nonendemic, the last three columns
containing the probability of each endemicity status (and summing to 1).
}
\description{
Extract the probabilities of each endemicity status for tip and internal node
states from the output of an optimisation performed with BioGeoBEARS
}
