% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clade_name.R
\name{extract_clade_name}
\alias{extract_clade_name}
\title{Creates a name for a clade depending on whether all the species of the clade
have the same genus name or whether the clade is composed of multiple genera,
in which case it will create a unique clade name by concatinating the genus
names}
\usage{
extract_clade_name(clade)
}
\arguments{
\item{clade}{A numeric vector which the indices of the species which are
in the island clade.}
}
\value{
Character
}
\description{
Creates a name for a clade depending on whether all the species of the clade
have the same genus name or whether the clade is composed of multiple genera,
in which case it will create a unique clade name by concatinating the genus
names
}
