% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_endemic_singleton.R
\name{extract_endemic_singleton}
\alias{extract_endemic_singleton}
\title{Extracts the information for an endemic species from a phylogeny
(specifically `phylo4d`  object from `phylobase` package) and stores it in
in an `island_colonist` class}
\usage{
extract_endemic_singleton(phylod, species_label)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}
}
\value{
An object of `island_colonist` class
}
\description{
Extracts the information for an endemic species from a phylogeny
(specifically `phylo4d`  object from `phylobase` package) and stores it in
in an `island_colonist` class
}
\examples{
set.seed(
  1,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
phylo <- ape::rcoal(10)
phylo <- phylobase::phylo4(phylo)
endemicity_status <- sample(c("not_present", "endemic", "nonendemic"),
                              size = length(phylobase::tipLabels(phylo)),
                              replace = TRUE)
phylod <- phylobase::phylo4d(phylo, as.data.frame(endemicity_status))
extract_endemic_singleton(phylod = phylod, species_label = "t7")
}
