% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_multi_tip_endemic.R
\name{extract_multi_tip_endemic}
\alias{extract_multi_tip_endemic}
\title{Extracts the information for an endemic species which has multiple tips in
the phylogeny (i.e. more than one sample per species) from a phylogeny
(specifically `phylo4d`  object from `phylobase` package)
and stores it in an `island_colonist` class}
\usage{
extract_multi_tip_endemic(phylod, species_label)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}
}
\value{
An object of `Island_colonist` class
}
\description{
Extracts the information for an endemic species which has multiple tips in
the phylogeny (i.e. more than one sample per species) from a phylogeny
(specifically `phylo4d`  object from `phylobase` package)
and stores it in an `island_colonist` class
}
