% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nonendemic.R
\name{extract_nonendemic}
\alias{extract_nonendemic}
\title{Extracts the information for a non-endemic species from a phylogeny
(specifically `phylo4d`  object from `phylobase` package) and stores it in
in an `island_colonist` class}
\usage{
extract_nonendemic(phylod, species_label)
}
\value{

}
\description{
Extracts the information for a non-endemic species from a phylogeny
(specifically `phylo4d`  object from `phylobase` package) and stores it in
in an `island_colonist` class
}
