% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nonendemic.R
\name{extract_nonendemic_forced}
\alias{extract_nonendemic_forced}
\title{Extract non-endemic colonist that is forced to be a singleton by user}
\usage{
extract_nonendemic_forced(phylod, species_label, island_tbl)
}
\arguments{
\item{phylod}{A \code{phylo4d} object from the package \code{phylobase} containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}

\item{island_tbl}{An instance of the \code{Island_tbl} class.}
}
\value{
An object of \code{phylo4d} class with tip and node data
}
\description{
Extract non-endemic colonist that is forced to be a singleton by user
}
\keyword{internal}
