% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_species_min.R
\name{extract_species_min}
\alias{extract_species_min}
\title{Title}
\usage{
extract_species_min(phylod, species_label, species_endemicity, island_tbl)
}
\arguments{
\item{island_tbl}{}
}
\value{

}
\description{
Title
}
