% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stem_age.R
\name{extract_stem_age}
\alias{extract_stem_age}
\title{Extracts the stem age from the phylogeny when the a species is known to
belong to a genus but is not itself in the phylogeny and there are members
of the same genus are in the phylogeny. Extraction method can either be
'min' or 'asr' as in extract_island_species(). When extraction_method = 'asr'
the constrain_to_island is ignored and the reconstructed node states are used
to determine the stem age and not the stem age of the genus or stem age of
island species within a genus (i.e. constrain_to_island = TRUE).}
\usage{
extract_stem_age(
  genus_name,
  phylod,
  extraction_method,
  constrain_to_island = FALSE
)
}
\arguments{
\item{genus_name}{Character string of genus name to be matched with a genus name from
the tip labels in the phylogeny}

\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{extraction_method}{A character string specifying whether the
colonisation time extracted is the minimum time (`min`) (before the present),
or the most probable time under ancestral state reconstruction (`asr`).}

\item{constrain_to_island}{Boolean determining whether the stem age extracted
is constrain to the species in the genus that are on the island, (i.e. stem
age of the island species). If FALSE the stem age of the genus is found
independent of whether its constituent species are on the island or not}
}
\value{
Numeric
}
\description{
Extracts the stem age from the phylogeny when the a species is known to
belong to a genus but is not itself in the phylogeny and there are members
of the same genus are in the phylogeny. Extraction method can either be
'min' or 'asr' as in extract_island_species(). When extraction_method = 'asr'
the constrain_to_island is ignored and the reconstructed node states are used
to determine the stem age and not the stem age of the genus or stem age of
island species within a genus (i.e. constrain_to_island = TRUE).
}
\examples{
# In this example the parrot clade is the genus of interest only the parrots
# are endemic to the island and all the passerines are not on the island
set.seed(1)
tree <- ape::rcoal(10)
tree$tip.label <- c(
  "passerine_a", "passerine_b", "passerine_c", "passerine_d", "passerine_e",
  "passerine_f", "parrot_a", "parrot_b", "parrot_c", "passerine_j")
tree <- phylobase::phylo4(tree)
endemicity_status <- c(
  "not_present", "not_present", "not_present", "not_present", "not_present",
  "not_present", "endemic", "endemic", "endemic", "not_present")
phylod <- phylobase::phylo4d(tree, as.data.frame(endemicity_status))
DAISIEprep::plot_phylod(phylod)
# the species 'parrot_a' is removed and becomes the missing species we want
# to the know the stem age for
phylod <- phylobase::subset(x = phylod, tips.exclude = "parrot_a")
DAISIEprep::plot_phylod(phylod)
# here we set constrain_to_island = TRUE so it only looks at species in that
# genus that are on the island and constrains the stem age to the stem of the
# island subclade
extract_stem_age(
  genus_name = "parrot",
  phylod = phylod,
  extraction_method = "min",
  constrain_to_island = TRUE
)
# here we set constrain_to_island = FALSE so it extracts the stem age of the
# genus independent of whether any species in that genus are on the island
# or not
extract_stem_age(
  genus_name = "parrot",
  phylod = phylod,
  extraction_method = "min",
  constrain_to_island = FALSE
)
# here we use the extraction_method = "asr" which requires ancestral node
# states in the tree. When "asr" is used the constrain_to_island argument
# becomes redundant and is ignored, setting constrain_to_island to TRUE or
# FALSE will not change the stem age extracted
phylod <- add_asr_node_states(
  phylod = phylod,
  asr_method = "parsimony",
  tie_preference = "mainland"
)
DAISIEprep::plot_phylod(phylod)
extract_stem_age(
  genus_name = "parrot",
  phylod = phylod,
  extraction_method = "asr",
  constrain_to_island = FALSE
)
}
