% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stem_age_min.R
\name{extract_stem_age_min}
\alias{extract_stem_age_min}
\title{Extracts the stem age from the phylogeny when the a species is known to
belong to a genus but is not itself in the phylogeny and there are members
of the same genus are in the phylogeny using the 'min' extraction method}
\usage{
extract_stem_age_min(genus_in_tree, phylod, constrain_to_island)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}
}
\value{
Numeric
}
\description{
Extracts the stem age from the phylogeny when the a species is known to
belong to a genus but is not itself in the phylogeny and there are members
of the same genus are in the phylogeny using the 'min' extraction method
}
