% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sensitivity.R
\name{format_sensitivity}
\alias{format_sensitivity}
\title{Formats sensitivity data into tidy data}
\usage{
format_sensitivity(sensitivity_data)
}
\value{
tibble
}
\description{
Formats sensitivity data into tidy data
}
\keyword{internal}
