% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GalapagosTrees}
\alias{GalapagosTrees}
\alias{coccyzus_tree}
\alias{columbiformes_tree}
\alias{finches_tree}
\alias{mimus_tree}
\alias{myiarchus_tree}
\alias{progne_tree}
\alias{pyrocephalus_tree}
\alias{setophaga_tree}
\title{Phylogenetic trees of the Galapagos bird lineages and sister species on the
mainland.}
\format{
A \code{phylo} object (from the \pkg{ape} package)

An object of class \code{phylo} of length 5.

An object of class \code{phylo} of length 4.

An object of class \code{phylo} of length 5.

An object of class \code{phylo} of length 4.

An object of class \code{phylo} of length 5.

An object of class \code{phylo} of length 5.

An object of class \code{phylo} of length 5.
}
\source{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015). Equilibrium
and non-equilibrium dynamics simultaneously operate in the Galapagos
islands. Ecology Letters 18: 844-852.
}
\usage{
coccyzus_tree

columbiformes_tree

finches_tree

mimus_tree

myiarchus_tree

progne_tree

pyrocephalus_tree

setophaga_tree
}
\description{
Each dataset is a phylogenetic tree used to extract the Galapagos avifaunal
lineages.
}
\keyword{datasets}
