% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endemic_species.R
\name{get_endemic_species}
\alias{get_endemic_species}
\title{Checks whether the focal species (given by its tip lable in species_label
argument) is part of an endemic clade on the island and a vector of the
endemic species, either a single species for a singleton or multiple species
in an endemic clade.}
\usage{
get_endemic_species(phylod, species_label)
}
\arguments{
\item{phylod}{A \code{phylo4d} object from the package \code{phylobase} containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}
}
\value{
Named numeric vector
}
\description{
Checks whether the focal species (given by its tip lable in species_label
argument) is part of an endemic clade on the island and a vector of the
endemic species, either a single species for a singleton or multiple species
in an endemic clade.
}
\keyword{internal}
