% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_colonist-accessors.R
\name{get_missing_species,island_colonist-method}
\alias{get_missing_species,island_colonist-method}
\title{Title}
\usage{
\S4method{get_missing_species}{island_colonist}(x)
}
\arguments{
\item{island_colonist}{}
}
\value{

}
\description{
Title
}
