% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemicity_to_states.R
\name{get_tip_states}
\alias{get_tip_states}
\title{Extract tip states from a phylod object}
\usage{
get_tip_states(phylod)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}
}
\value{
an integer vector of tip states, as expected by SSE models
}
\description{
Extract tip states from a phylod object
}
