% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_island_clade.R
\name{is_island_clade}
\alias{is_island_clade}
\title{Checks whether the focal species (given by its tip lable in species_label
argument) is part of an endemic clade on the island}
\usage{
is_island_clade(phylod, species_label)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}
}
\value{
Boolean
}
\description{
Checks whether the focal species (given by its tip lable in species_label
argument) is part of an endemic clade on the island
}
\examples{
x = 1
}
