% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_colonist-accessors.R
\name{get_clade_name}
\alias{get_clade_name}
\alias{get_clade_name,island_colonist-method}
\title{Gets clade name from object of `island_colonist` class.}
\usage{
get_clade_name(x)

\S4method{get_clade_name}{island_colonist}(x)
}
\arguments{
\item{x}{An object whose class is determined by the signature.}
}
\value{
Character string
}
\description{
Gets clade name from object of `island_colonist` class.
}
