% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_colonist-class.R
\docType{class}
\name{Island_colonist-class}
\alias{Island_colonist-class}
\title{Defines the `island_tbl` class which is used when extracting information
from the phylogenetic and island data to be used for constructing a
`daisie_data_tbl`}
\description{
Defines the `island_tbl` class which is used when extracting information
from the phylogenetic and island data to be used for constructing a
`daisie_data_tbl`
}
\section{Slots}{

\describe{
\item{\code{clade_name}}{character.}

\item{\code{status}}{character.}

\item{\code{missing_species}}{character.}

\item{\code{col_time}}{numeric.}

\item{\code{col_max_age}}{logical.}

\item{\code{branching_times}}{numeric.}

\item{\code{min_age}}{numeric.}

\item{\code{species}}{character.}

\item{\code{clade_type}}{numeric.}
}}

