% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_colonist-class.R
\name{island_colonist}
\alias{island_colonist}
\title{Constructor for Island_colonist}
\usage{
island_colonist(
  clade_name = NA_character_,
  status = NA_character_,
  missing_species = NA_real_,
  branching_times = NA_real_
)
}
\arguments{
\item{clade_name}{Character name of the colonising clade.}

\item{status}{Character endemicity status of the colonising clade.}

\item{missing_species}{Numeric number of missing species from the phylogeny
that belong to the colonising clade.}

\item{branching_times}{Numeric vector of one or more elements where the first
element is the colonisation time and subsequent elements are the branching
times on the island.}
}
\value{
Object of `Island_colonist` class.
}
\description{
Constructor for Island_colonist
}
\examples{
# Without initial values
colonist <- island_colonist()

# With initial values
colonist <- island_colonist(
  clade_name = "bird",
  status = "endemic",
  missing_species = 0,
  branching_times = 0.5
)
}
