% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_tbl-accessors.R
\name{get_island_tbl}
\alias{get_island_tbl}
\alias{get_island_tbl,Island_tbl-method}
\alias{set_island_tbl<-}
\alias{set_island_tbl<-,Island_tbl-method}
\alias{get_extracted_species}
\alias{get_extracted_species,Island_tbl-method}
\alias{set_extracted_species<-}
\alias{set_extracted_species<-,Island_tbl-method}
\alias{get_num_phylo_used}
\alias{get_num_phylo_used,Island_tbl-method}
\alias{set_num_phylo_used<-}
\alias{set_num_phylo_used<-,Island_tbl-method}
\title{Accessor functions for the data (slots) in objects of the
\code{\linkS4class{Island_tbl}} class}
\usage{
get_island_tbl(x)

\S4method{get_island_tbl}{Island_tbl}(x)

set_island_tbl(x) <- value

\S4method{set_island_tbl}{Island_tbl}(x) <- value

get_extracted_species(x)

\S4method{get_extracted_species}{Island_tbl}(x)

set_extracted_species(x) <- value

\S4method{set_extracted_species}{Island_tbl}(x) <- value

get_num_phylo_used(x)

\S4method{get_num_phylo_used}{Island_tbl}(x)

set_num_phylo_used(x) <- value

\S4method{set_num_phylo_used}{Island_tbl}(x) <- value
}
\arguments{
\item{x}{An object whose class is determined by the signature.}

\item{value}{A value which can take several forms to be assigned to an object
of a class.}
}
\value{
Getter function (get_*) returns a data frame, the setter function
(set_*) returns the modified Island_tbl class.
}
\description{
Accessor functions for the data (slots) in objects of the
\code{\linkS4class{Island_tbl}} class
}
\examples{
island_tbl <- island_tbl()
get_island_tbl(island_tbl)
set_island_tbl(island_tbl) <- data.frame(
  clade_name = "birds",
  status = "endemic",
  missing_species = 0,
  branching_times = I(list(c(1.0, 0.5)))
)
}
\author{
Joshua W. Lambert
}
