% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_tbl-class.R
\docType{class}
\name{Island_tbl-class}
\alias{Island_tbl-class}
\title{Defines the `island_tbl` class which is used when extracting information
from the phylogenetic and island data to be used for constructing a
`daisie_data_tbl`}
\description{
Defines the `island_tbl` class which is used when extracting information
from the phylogenetic and island data to be used for constructing a
`daisie_data_tbl`
}
\section{Slots}{

\describe{
\item{\code{island_tbl}}{data frame.}

\item{\code{metadata}}{list.}
}}

