% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_tbl-class.R
\name{island_tbl}
\alias{island_tbl}
\title{Constructor function for `Island_tbl` class}
\usage{
island_tbl()
}
\value{
An Island_tbl object.
}
\description{
Constructor function for `Island_tbl` class
}
