% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mimus_phylod}
\alias{mimus_phylod}
\title{A phylogenetic tree of mimus species with endemicity status as tip states.}
\format{
\subsection{\code{mimus_phylod}}{

A \code{phylo4d} object (from the \pkg{phylobase} package) with
29 tips and
28 internal
nodes.
}
}
\source{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015). Equilibrium
and non-equilibrium dynamics simultaneously operate in the Galapagos
islands. Ecology Letters 18: 844-852.
}
\usage{
mimus_phylod
}
\description{
A phylogenetic tree of mimus species with endemicity status as tip states.
}
\keyword{datasets}
