% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_island_tbl-class.R
\docType{class}
\name{Multi_island_tbl-class}
\alias{Multi_island_tbl-class}
\title{Defines the `Multi_island_tbl` class which is multiple `Island_tbl`s.}
\description{
Defines the `Multi_island_tbl` class which is multiple `Island_tbl`s.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a list of `Island_tbl`.}
}}

