% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myiarchus_phylod}
\alias{myiarchus_phylod}
\title{A phylogenetic tree of myiarchus species with endemicity status as tip states.}
\format{
\subsection{\code{myiarchus_phylod}}{

A \code{phylo4d} object (from the \pkg{phylobase} package) with
13 tips and
12 internal
nodes.
}
}
\source{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015). Equilibrium
and non-equilibrium dynamics simultaneously operate in the Galapagos
islands. Ecology Letters 18: 844-852.
}
\usage{
myiarchus_phylod
}
\description{
A phylogenetic tree of myiarchus species with endemicity status as tip states.
}
\keyword{datasets}
