% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_phylo}
\alias{plant_phylo}
\title{A phylogenetic tree of plant species.}
\format{
\subsection{\code{plant_phylo}}{

A \code{phylo} object (from the \pkg{ape} package) with 10 tips and 9 internal
nodes. Each tip label is formatted \code{"Plant_*"} with the letters \code{a}, \code{b},
etc. used as species names.
}
}
\usage{
plant_phylo
}
\description{
This is a mock dataset for demonstrating the \pkg{DAISIEprep} package.
}
\keyword{datasets}
