% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{plot_performance}
\alias{plot_performance}
\title{Plots performance results for a grouping variable (prob_on_island or
prob_endemic).}
\usage{
plot_performance(performance_data, group_by)
}
\arguments{
\item{performance_data}{Tibble of collated performance results}

\item{group_by}{A variable to partition by for plotting. Uses data masking
so variable does not need to be quoted.}
}
\value{
ggplot2 object
}
\description{
Plots performance results for a grouping variable (prob_on_island or
prob_endemic).
}
