% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pyrocephalus_phylod}
\alias{pyrocephalus_phylod}
\title{A phylogenetic tree of pyrocephalus species with endemicity status as
tip states.}
\format{
\subsection{\code{pyrocephalus_phylod}}{

A \code{phylo4d} object (from the \pkg{phylobase} package) with
66 tips and
65 internal
nodes.
}
}
\source{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015). Equilibrium
and non-equilibrium dynamics simultaneously operate in the Galapagos
islands. Ecology Letters 18: 844-852.
}
\usage{
pyrocephalus_phylod
}
\description{
A phylogenetic tree of pyrocephalus species with endemicity status as
tip states.
}
\keyword{datasets}
