% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_performance.R
\name{read_performance}
\alias{read_performance}
\title{Reads in performance analysis results from inst/performance_data and formats
the data ready for plotting}
\usage{
read_performance()
}
\value{
List of two data frames
}
\description{
Reads in performance analysis results from inst/performance_data and formats
the data ready for plotting
}
\keyword{internal}
