% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sensitivity.R
\name{read_sensitivity}
\alias{read_sensitivity}
\title{Reads in the results from the sensitivity analysis saved in the
inst/sensitivity_data folder}
\usage{
read_sensitivity()
}
\value{
List of of lists with parameter estimates for the mammal DNA-only
phylogeny, mammal complete phylogeny and the parameter set used
}
\description{
Reads in the results from the sensitivity analysis saved in the
inst/sensitivity_data folder
}
\keyword{internal}
