% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_up}
\alias{round_up}
\title{Rounds numbers using the round up method, rather than the round to the
nearest even number method used by the base function \code{round}.}
\usage{
round_up(n, digits = 0)
}
\arguments{
\item{n}{A numeric to be rounded.}

\item{digits}{A numeric specifying which decimal places to round to}
}
\value{
Numeric
}
\description{
Rounds numbers using the round up method, rather than the round to the
nearest even number method used by the base function \code{round}.
}
