% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_colonist-accessors.R
\name{set_clade_name<-}
\alias{set_clade_name<-}
\title{Sts clade name of object of `island_colonist` class.}
\usage{
set_clade_name(x) <- value
}
\arguments{
\item{x}{An object whose class is determined by the signature.}

\item{value}{A value which can take several forms to be assigned to an object
of a class.}
}
\value{
Nothing
}
\description{
Sts clade name of object of `island_colonist` class.
}
