% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_tbl-accessors.R
\name{set_island_tbl<-}
\alias{set_island_tbl<-}
\title{Sets value for `island_tbl` class.}
\usage{
set_island_tbl(x) <- value
}
\arguments{
\item{x}{An object whose class is determined by the signature.}

\item{value}{A value which can take several forms to be assigned to an object
of a class.}
}
\value{
Nothing
}
\description{
Sets value for `island_tbl` class.
}
