% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemicity_to_states.R
\name{states_to_endemicity}
\alias{states_to_endemicity}
\title{Convert SSE states back to endemicity status}
\usage{
states_to_endemicity(states)
}
\arguments{
\item{states}{integer vector of tip states, as expected by SSE models}
}
\value{
character vector with values "endemic", "nonendemic" and/or
"not_present"
}
\description{
Convert SSE states back to endemicity status
}
