% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_biogeobears_input.R
\name{write_biogeobears_input}
\alias{write_biogeobears_input}
\title{Write input files for BioGeoBEARS}
\usage{
write_biogeobears_input(phylod, path_to_phylo, path_to_biogeo)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{path_to_phylo}{string specifying the path and name to write the phylogeny file to.}

\item{path_to_biogeo}{string specifying the path and name to write the biogeography file to.}
}
\description{
Write input files for a BioGeoBEARS analysis, i.e. a phlyogenetic tree in
Newick format and occurrence data in PHYLIP format.
}
