% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_biogeobears_input.R
\name{write_newick_file}
\alias{write_newick_file}
\title{Write tree input file for BioGeoBEARS}
\usage{
write_newick_file(phylod, path_to_phylo)
}
\arguments{
\item{phylod}{A \code{phylo4d} object from the package \code{phylobase} containing
phylogenetic and endemicity data for each species.}

\item{path_to_phylo}{string specifying the path and name to write the file to.}
}
\value{
Nothing, called for side-effects.
}
\description{
Write a text file containing a phylogenetic tree in the Newick format
expected by BioGeoBEARS
}
