% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_biogeobears_input.R
\name{write_phylip_biogeo_file}
\alias{write_phylip_biogeo_file}
\title{Write biogeography input file for BioGeoBEARS}
\usage{
write_phylip_biogeo_file(phylod, path_to_biogeo)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{path_to_biogeo}{string specifying the path and name to write the file to.}
}
\description{
Write a text file containing occurrence data for all tips in the PHYLIP format
expected by BioGeoBEARS
}
