% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outgroup.R
\name{add_outgroup}
\alias{add_outgroup}
\title{Add an outgroup species to a given phylogeny.}
\usage{
add_outgroup(phylo)
}
\arguments{
\item{phylo}{A phylogeny either as a `phylo` (from the `ape` package) or
`phylo4` (from the `phylobase` package) object.}
}
\value{
A `phylo` object
}
\description{
Add an outgroup species to a given phylogeny.
}
\examples{
phylo <- ape::rcoal(10)
phylo_with_outgroup <- add_outgroup(phylo)
}
