% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_daisie_datatable.R
\name{as_daisie_datatable}
\alias{as_daisie_datatable}
\title{Converts the `island_tbl` class to a data frame in the format of a DAISIE
data table (see DAISIE R package for details). This can then be input into
`DAISIEprep::create_daisie_data` function which creates the list input into
the DAISIE ML models.}
\usage{
as_daisie_datatable(island_tbl, island_age, col_uncertainty = NULL)
}
\arguments{
\item{island_tbl}{An instance of the `Island_tbl` class.}

\item{island_age}{Age of the island in appropriate units.}

\item{col_uncertainty}{A character, either "max", "min", or NULL. "max" makes
everything a max age colonisation, "min" applies min ages when available,
and NULL uses the precise times of colonisation, or the min ages when they
are informative.}
}
\value{
A data frame in the format of a DAISIE data table
}
\description{
Converts the `island_tbl` class to a data frame in the format of a DAISIE
data table (see DAISIE R package for details). This can then be input into
`DAISIEprep::create_daisie_data` function which creates the list input into
the DAISIE ML models.
}
\examples{
phylod <- create_test_phylod(10)
island_tbl <- extract_island_species(
  phylod = phylod,
  extraction_method = "asr"
)
daisie_datatable <- as_daisie_datatable(
  island_tbl = island_tbl,
  island_age = 0.2,
  col_uncertainty = "no"
)
}
