% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_colonist_to_tbl.R
\name{bind_colonist_to_tbl}
\alias{bind_colonist_to_tbl}
\title{Takes an existing instance of an \code{Island_tbl} class and bind the information
from the instance of an \code{Island_colonist} class to it}
\usage{
bind_colonist_to_tbl(island_colonist, island_tbl)
}
\arguments{
\item{island_colonist}{An instance of the \code{Island_colonist} class.}

\item{island_tbl}{An instance of the \code{Island_tbl} class.}
}
\value{
An object of \code{Island_tbl} class
}
\description{
Takes an existing instance of an \code{Island_tbl} class and bind the information
from the instance of an \code{Island_colonist} class to it
}
\examples{
island_colonist <- DAISIEprep::island_colonist(
  clade_name = "bird",
  status = "endemic",
  missing_species = 0,
  col_time = 1,
  col_max_age = FALSE,
  branching_times = 0.5,
  species = "bird_a",
  clade_type = 1
)
island_tbl <- island_tbl()
bind_colonist_to_tbl(
  island_colonist = island_colonist,
  island_tbl = island_tbl
)
}
