% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_colonist-class.R
\name{check_island_colonist}
\alias{check_island_colonist}
\title{Checks the validity of the Island_colonist class}
\usage{
check_island_colonist(object)
}
\arguments{
\item{object}{Instance of the island_colonist class}
}
\value{
Boolean or errors
}
\description{
Checks the validity of the Island_colonist class
}
\examples{
island_col <- island_colonist()
check_island_colonist(island_col)
}
