% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_tbl-class.R
\name{check_island_tbl}
\alias{check_island_tbl}
\title{Checks the validity of the Island_tbl class}
\usage{
check_island_tbl(object)
}
\arguments{
\item{object}{Instance of the Island_tbl class}
}
\value{
Boolean or errors
}
\description{
Checks the validity of the Island_tbl class
}
\examples{
tbl <- island_tbl()
check_island_tbl(tbl)
}
