% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_phylod.R
\name{create_test_phylod}
\alias{create_test_phylod}
\title{Creates phylod objects to be used in tests}
\usage{
create_test_phylod(test_scenario)
}
\value{
A `phylo4d` object
}
\description{
Creates phylod objects to be used in tests
}
\examples{
create_test_phylod(test_scenario = 1)
}
