% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for function in the DAISIEprep package}
\usage{
default_params_doc(
  island_colonist,
  island_tbl,
  phylod,
  species_label,
  x,
  value
)
}
\arguments{
\item{island_colonist}{An instance of the `island_colonist` class.}

\item{island_tbl}{An instance of the `island_tbl` class.}

\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}

\item{x}{An object whose class is determined by the signature.}

\item{value}{A value which can take several forms to be assigned to an object
of a class.}
}
\value{
Nothing
}
\description{
Documentation for function in the DAISIEprep package
}
\author{
Joshua W. Lambert
}
