% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for function in the DAISIEprep package}
\usage{
default_params_doc(
  island_colonist,
  island_tbl,
  phylod,
  extraction_method,
  species_label,
  species_endemicity,
  x,
  value,
  clade_name,
  status,
  missing_species,
  branching_times,
  min_age,
  endemic_clade,
  phylo,
  island_species,
  descendants,
  clade,
  asr_method,
  tie_preference,
  earliest_col,
  include_not_present,
  missing_species_df,
  node_pies,
  daisie_datatable,
  island_age,
  num_mainland_species,
  num_clade_types,
  list_type2_clades,
  prop_type2_pool,
  epss,
  verbose
)
}
\arguments{
\item{island_colonist}{An instance of the `Island_colonist` class.}

\item{island_tbl}{An instance of the `Island_tbl` class.}

\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{extraction_method}{A character string specifying whether the
colonisation time extracted is the minimum time (`min`) (before the present),
or the most probable time under ancestral state reconstruction (`asr`).}

\item{species_label}{The tip label of the species of interest.}

\item{species_endemicity}{A character string with the endemicity, either
"endemic" or "nonendemic" of an island species, or "not_present" if not on
the island.}

\item{x}{An object whose class is determined by the signature.}

\item{value}{A value which can take several forms to be assigned to an object
of a class.}

\item{clade_name}{Character name of the colonising clade.}

\item{status}{Character endemicity status of the colonising clade.}

\item{missing_species}{Numeric number of missing species from the phylogeny
that belong to the colonising clade.}

\item{branching_times}{Numeric vector of one or more elements where the first
element is the colonisation time and subsequent elements are the branching
times on the island.}

\item{min_age}{Numeric minimum age (time before the present) that the species
must have colonised the island by. This is known when there is a branching
on the island, either in species or subspecies.}

\item{endemic_clade}{Named vector with all the species from a clade.}

\item{phylo}{A `phylo4` object from the package `phylobase`.}

\item{island_species}{Data frame with two columns. The first is a character
string of the tip_labels with the tip names of the species on the island.
The second column a character string of the endemicity status of the species,
either endemic or nonendemic.}

\item{descendants}{A vector character strings with the names of species to
determine whether they are the same species.}

\item{clade}{A numeric vector which the indices of the species which are
in the island clade.}

\item{asr_method}{A character string, either "parsimony" or "mk" determines
whether a maximum parsimony or continuous-time markov model reconstructs the
ancestral states at each node. See documentation in
`castor::asr_maximum_parsimony` or `castor::asr_mk` in `castor` R package
for details on the methods used.}

\item{tie_preference}{Character string, either "island" or "mainland" to
choose the most probable state at each node using the `max.col()` function.
When a node has island presence and absence equally probable we need to
decide whether that species should be considered on the island. To consider
it on the island use `ties.method = "last"` in the `max.col()` function, if
you consider it not on the island use `ties.method = "first"`. Default is
"island".}

\item{earliest_col}{A boolean to determine whether to take the colonisation
time as the most probable time (FALSE) or the earliest possible colonisation
time (TRUE), where the probability of a species being on the island is
non-zero. Default is FALSE.}

\item{include_not_present}{A boolean determining whether species not present
on the island should be included in island colonist when embedded within an
island clade. Default is FALSE.}

\item{missing_species_df}{A data frame containing the names of the clades
extracted from the phylogeny and the number of missing species in each clade.
Not every clade extracted from the phylogeny needs to be included if the
number of missing species in them clades is zero.}

\item{node_pies}{Boolean determining if pie charts of the probabilities of
a species being present on the island. If TRUE the correct data is required
in the phylod object.}

\item{daisie_datatable}{A data frame where each row on the table represents
an independent colonisation event. The table has the following four columns:
* Clade_name: name of independent colonisation event
* Status: One of the following categories:
- "Non_endemic": applies to non-endemic species when an approximate
colonisation time is known.
- "Non_endemic_MaxAge": applies to non-endemic species for cases where
colonisation time is unknown.
- "Endemic": applies to endemic species or endemic clades when an approximate
colonisation time is known.
- "Endemic_MaxAge": appies to endemic species or endemic clades where the
colonisation time is unknown, or when the user wants to specify an upper
bound for colonisation. This could for example apply to endemic species that
have recently gone extinct because of anthropogenic causes, and which are
not included in the phylogeny ("NA" should be given in the branching times
column). It could also apply to insular radiations with long stem branches,
for which the time of the first cladogenetic event is known, but the precise
time of the colonisation is not.
- "Endemic&Non_Endemic": where endemic clade is present and its mainland
ancestor has re-colonised.
* Missing_species: Number of island species that were not sampled for a
particular clade (only applicable for "Endemic" clades). If NA is given in
branching times column, this should be equal to the number of species in the
clade minus 1.
* Branching_times: Stem age of the population/species in the case of
"Non_endemic", "Non_endemic_MaxAge", and "Endemic" species with no extant
clade relatives on the island. Set "NA" if colonisation time is unknown and
no upper bound is known. For "Endemic" cladogenetic species these should be
branching times of the radiation, including the stem age of the radiation
(colonisaton time estimate).}

\item{island_age}{Age of the island in appropriate units.}

\item{num_mainland_species}{The size of the mainland pool, i.e. the number
of species that can potentially colonise the island.}

\item{num_clade_types}{Number of clade types. Default num_clade_types = 1 all
species are considered to belong to the same macroevolutionary process. If
num_clade_types = 2, there are two types of clades with distinct
macroevolutionary processes.}

\item{list_type2_clades}{If num_clade_types = 2, list_type2_clades specifies
the names of the clades that have a distinct macroevolutionary process. The
names must match those in the "Clade_name" column of the source data table.
If num_clade_types = 1, then list_type2_clades = NA should be specified
(default).}

\item{prop_type2_pool}{Specifies the fraction of potential mainland colonists
that have a distinct macroevolutionary process. Applies only if
number_clade_types = 2. Default "proportional" sets the fraction to be
proportional to the number of clades of distinct macroevolutionary process
that have colonised the island. Alternatively, the user can specify a value
between 0 and 1 (e.g. if the mainland pool size is 1000 and prop_type2_pool
= 0.02 then the number of type 2 species is 20).}

\item{epss}{Default = 1e-5 should be appropriate in most cases. This value
is used to set the maximum age of colonisation of "Non_endemic_MaxAge" and
"Endemic_MaxAge" species to an age that is slightly younger than the island
for cases when the age provided for that species is older than the island.
The new maximum age is then used as an upper bound to integrate over all
possible colonisation times.}

\item{verbose}{Boolean. States if intermediate results should be printed to
console. Defaults to TRUE.}
}
\value{
Nothing
}
\description{
Documentation for function in the DAISIEprep package
}
\author{
Joshua W. Lambert
}
