% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_asr_clade.R
\name{extract_asr_clade}
\alias{extract_asr_clade}
\title{Extracts an island clade based on the ancestral state reconstruction of the
species presence on the island, therefore this clade can contain
non-endemic species as well as endemic species.}
\usage{
extract_asr_clade(phylod, species_label, ancestor, clade)
}
\arguments{
\item{clade}{}
}
\value{

}
\description{
Extracts an island clade based on the ancestral state reconstruction of the
species presence on the island, therefore this clade can contain
non-endemic species as well as endemic species.
}
