% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_multi_tip_species.R
\name{extract_multi_tip_species}
\alias{extract_multi_tip_species}
\title{Extracts the information for a species (endemic or non-endemic) which has
multiple tips in the phylogeny (i.e. more than one sample per species) from
a phylogeny (specifically `phylo4d`  object from `phylobase` package)
and stores it in an `island_colonist` class}
\usage{
extract_multi_tip_species(phylod, species_label, species_endemicity)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}

\item{species_endemicity}{A character string with the endemicity, either
"endemic" or "nonendemic" of an island species, or "not_present" if not on
the island.}
}
\value{
An object of `Island_colonist` class
}
\description{
Extracts the information for a species (endemic or non-endemic) which has
multiple tips in the phylogeny (i.e. more than one sample per species) from
a phylogeny (specifically `phylo4d`  object from `phylobase` package)
and stores it in an `island_colonist` class
}
