% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stem_age_min.R
\name{extract_stem_age_min}
\alias{extract_stem_age_min}
\title{Extracts the stem age from the phylogeny when the a species is known to
belong to a genus but is not itself in the phylogeny and there are members
of the same genus are in the phylogeny using the 'min' extraction method}
\usage{
extract_stem_age_min(genus_in_tree, phylod, constrain_to_island)
}
\arguments{
\item{genus_in_tree}{A numeric vector that indicates which species in the
genus are in the tree}

\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{constrain_to_island}{Boolean determining whether the stem age extracted
is constrain to the species in the genus that are on the island, (i.e. stem
age of the island species). If FALSE the stem age of the genus is found
independent of whether its constituent species are on the island or not}
}
\value{
Numeric
}
\description{
Extracts the stem age from the phylogeny when the a species is known to
belong to a genus but is not itself in the phylogeny and there are members
of the same genus are in the phylogeny using the 'min' extraction method
}
