% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_multi_tip_species.R
\name{is_multi_tip_species}
\alias{is_multi_tip_species}
\title{Checks if a non-endemic species is represented in the tree as multiple tips}
\usage{
is_multi_tip_species(phylod, species_label)
}
\arguments{
\item{phylod}{A `phylo4d` object from the package `phylobase` containing
phylogenetic and endemicity data for each species.}

\item{species_label}{The tip label of the species of interest.}
}
\value{
Boolean
}
\description{
Checks if a non-endemic species is represented in the tree as multiple tips
}
\keyword{internal}
