% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_multi_tip_species.R
\name{is_multi_tip_species}
\alias{is_multi_tip_species}
\title{Checks if a non-endemic species is represented in the tree as multiple tips}
\usage{
is_multi_tip_species(phylod, species_label)
}
\arguments{
\item{species_label}{}
}
\value{

}
\description{
Checks if a non-endemic species is represented in the tree as multiple tips
}
