% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_tbl-accessors.R
\name{get_island_tbl}
\alias{get_island_tbl}
\alias{get_island_tbl,island_tbl-method}
\alias{set_island_tbl<-,island_tbl-method}
\title{Methods to get and set the elements of the `island_tbl` class.}
\usage{
get_island_tbl(x)

\S4method{get_island_tbl}{island_tbl}(x)

\S4method{set_island_tbl}{island_tbl}(x) <- value
}
\arguments{
\item{x}{An object whose class is determined by the signature.}

\item{value}{A value which can take several forms to be assigned to an object
of a class.}
}
\value{
Data frame
}
\description{
Methods to get and set the elements of the `island_tbl` class.
}
