% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_island_tbl-class.R
\docType{class}
\name{Multi_island_tbl-class}
\alias{Multi_island_tbl-class}
\title{Defines the \code{Multi_island_tbl} class which is multiple \code{Island_tbl}s.}
\description{
Defines the \code{Multi_island_tbl} class which is multiple \code{Island_tbl}s.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a list of \code{Island_tbl}.}
}}

