% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_island_tbl-class.R
\name{multi_island_tbl}
\alias{multi_island_tbl}
\title{Constructor function for \code{Multi_island_tbl} class}
\usage{
multi_island_tbl()
}
\value{
A Multi_island_tbl object.
}
\description{
Constructor function for \code{Multi_island_tbl} class
}
