% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_multi_missing_species.R
\name{rm_multi_missing_species}
\alias{rm_multi_missing_species}
\title{Loops through the genera that have missing species and removes the ones that
are found in the missing genus list which have phylogenetic data. This is
useful when wanting to know which missing species have not been assigned to
the island_tbl using `add_multi_missing_species()`.}
\usage{
rm_multi_missing_species(missing_species, missing_genus, island_tbl)
}
\arguments{
\item{missing_species}{Numeric number of missing species from the phylogeny
that belong to the colonising clade.}

\item{missing_genus}{A list of character vectors containing the genera in
each island clade}

\item{island_tbl}{An instance of the `Island_tbl` class.}
}
\value{
Data frame
}
\description{
Loops through the genera that have missing species and removes the ones that
are found in the missing genus list which have phylogenetic data. This is
useful when wanting to know which missing species have not been assigned to
the island_tbl using `add_multi_missing_species()`.
}
\examples{
phylod <- DAISIEprep:::create_test_phylod(test_scenario = 6)
island_tbl <- suppressWarnings(extract_island_species(
 phylod = phylod,
 extraction_method = "asr",
))
phylod <- DAISIEprep:::create_test_phylod(test_scenario = 7)
island_tbl <- suppressWarnings(extract_island_species(
 phylod = phylod,
 extraction_method = "asr",
 island_tbl = island_tbl
))
missing_species <- data.frame(
  clade_name = "bird",
  missing_species = 1,
  endemicity_status = "endemic"
)
missing_genus <- list("bird", character(0))
rm_missing_species <- rm_multi_missing_species(
  missing_species = missing_species,
  missing_genus = missing_genus,
  island_tbl = island_tbl
)
}
